+++ far_ws/src/follow_ahead_rl/data/person_trajectories_rl.pkl
@@ -1,0 +1,273 @@
+(lp1
+(dp2
+S'start_person'
+p3
+(dp4
+S'pos'
+p5
+(F0.28888158043799628
+F3.0370323014611489
+tp6
+sS'orientation'
+p7
+F1.637575593895455
+ssS'start_robot'
+p8
+(dp9
+g5
+(F0.32666359875810103
+F3.919990197948422
+tp10
+sg7
+F-1.5317784182184064
+ssS'points'
+p11
+(lp12
+(F0.28937468625265733
+F3.0378905605691946
+tp13
+a(F0.24866160409639143
+F3.2337461137676455
+tp14
+a(F0.11534829103262575
+F3.382946034573389
+tp15
+a(F-0.055929426216984497
+F3.48622218837192
+tp16
+a(F-0.2398382474325228
+F3.565054919930692
+tp17
+a(F-0.43157760842205856
+F3.6219844536927233
+tp18
+a(F-0.62856615097609736
+F3.6567881112552318
+tp19
+a(F-0.82819360931797692
+F3.6690718019884736
+tp20
+a(F-1.0280714439916161
+F3.6584174061598338
+tp21
+a(F-1.2244347423736177
+F3.6197617029622959
+tp22
+a(F-1.4159299278353619
+F3.5614972807461278
+tp23
+a(F-1.5989013874377391
+F3.4807005800973099
+tp24
+a(F-1.7708966847379066
+F3.3785939231190305
+tp25
+a(F-1.9295556041628801
+F3.2567383826484169
+tp26
+a(F-2.0728409331467899
+F3.116985477668297
+tp27
+a(F-2.1987527093115653
+F2.961556205301656
+tp28
+a(F-2.3058792191178639
+F2.7924929215187912
+tp29
+a(F-2.3927747245595881
+F2.6122960224546214
+tp30
+a(F-2.4584777134882638
+F2.4233150660408609
+tp31
+a(F-2.5022458641106073
+F2.2280061730298391
+tp32
+a(F-2.5236070514950515
+F2.029049473062853
+tp33
+a(F-2.5222442465447843
+F1.8290479122996168
+tp34
+a(F-2.4969918288944752
+F1.6306257147860961
+tp35
+a(F-2.4478479276031484
+F1.4366875570455258
+tp36
+a(F-2.3756430009562846
+F1.2501145563438456
+tp37
+a(F-2.2815755051293936
+F1.073486539312174
+tp38
+a(F-2.1672435432245707
+F0.90938130541582995
+tp39
+a(F-2.0343040584635106
+F0.75984537026867005
+tp40
+a(F-1.8849339743608762
+F0.62672301084679549
+tp41
+a(F-1.7229789987567381
+F0.50916720177600649
+tp42
+a(F-1.5511371766811684
+F0.40675504927044398
+tp43
+a(F-1.3752213052382818
+F0.3114086388647388
+tp44
+a(F-1.1994127785021753
+F0.21570157108514434
+tp45
+a(F-1.0237871866393888
+F0.11967081226522501
+tp46
+a(F-0.84833305250134206
+F0.023309690164487289
+tp47
+a(F-0.6730559108506653
+F-0.073376831541102513
+tp48
+a(F-0.49795595729342101
+F-0.17039194482920642
+tp49
+a(F-0.32302834574700956
+F-0.2677364154390901
+tp50
+a(F-0.14827606847086205
+F-0.36540075793333493
+tp51
+a(F0.026302963735068411
+F-0.46338575997541309
+tp52
+a(F0.19885274659305471
+F-0.56478050469573515
+tp53
+a(F0.35294084533018161
+F-0.69229494652977586
+tp54
+a(F0.45915675086928959
+F-0.86177303733893684
+tp55
+a(F0.48502609350740389
+F-1.0603074411418996
+tp56
+a(F0.4221236454215026
+F-1.2503674159065843
+tp57
+a(F0.28261118875793884
+F-1.3937028578817268
+tp58
+a(F0.094547603628628976
+F-1.4619586790218295
+tp59
+a(F-0.10451255139966138
+F-1.4419341846538145
+tp60
+a(F-0.27473289877162343
+F-1.3368596381822819
+tp61
+a(F-0.38214964870868201
+F-1.1679668229928946
+tp62
+a(F-0.41832624318816192
+F-0.97114129737684118
+tp63
+a(F-0.41321413713906191
+F-0.77112225295527537
+tp64
+a(F-0.38907128703924393
+F-0.5725722999805376
+tp65
+a(F-0.38498782728663372
+F-0.37244376131961782
+tp66
+a(F-0.44883199353895115
+F-0.18269103502843628
+tp67
+a(F-0.57709314541174206
+F-0.029228770256833507
+tp68
+a(F-0.74316152225470755
+F0.082483675993139272
+tp69
+a(F-0.92692064832293952
+F0.16165272477343881
+tp70
+a(F-1.1199573354630206
+F0.21468289734489862
+tp71
+a(F-1.3167737145570211
+F0.25041235729433353
+tp72
+a(F-1.5156795110302421
+F0.27317569068407965
+tp73
+a(F-1.7130227955496513
+F0.3068246995040137
+tp74
+a(F-1.8994968488009856
+F0.37926882134472967
+tp75
+a(F-2.0595540819159348
+F0.49919275667524571
+tp76
+a(F-2.1816193255973344
+F0.65773895199340049
+tp77
+a(F-2.2570532428674253
+F0.84315072678000147
+tp78
+a(F-2.2914084540631587
+F1.0402443206064536
+tp79
+a(F-2.2901080601706685
+F1.2402759233364733
+tp80
+a(F-2.2581175626386027
+F1.4377323251371426
+tp81
+a(F-2.2230665423714777
+F1.6347115056130548
+tp82
+a(F-2.2044264407292227
+F1.8338793583760156
+tp83
+a(F-2.2111583586086594
+F2.0338633918102982
+tp84
+a(F-2.2544638884044494
+F2.2292823387653629
+tp85
+a(F-2.3329932109944536
+F2.4132592193244706
+tp86
+a(F-2.4440424212717682
+F2.5798147934008959
+tp87
+a(F-2.5788438119073902
+F2.7277096775038459
+tp88
+a(F-2.7128022853661395
+F2.8764034112362866
+tp89
+a(F-2.8098771249136991
+F3.0513819954305057
+tp90
+a(F-2.82572863079309
+F3.2507587681632857
+tp91
+a(F-2.7533478362498678
+F3.4372295209252619
+tp92
+a(F-2.6066171635280737
+F3.5733780018207364
+tp93
+asS'name'
+p94
+S'smoothed_one'
+p95
+sa.
