+++ far_ws/src/follow_ahead_rl/scripts/d4pg/configs/LunarLanderContinuous_d4pg.yml
@@ -1,0 +1,53 @@
+# Environment parameters
+
+env: LunarLanderContinuous-v2
+state_dim: 8
+action_dim: 2
+action_low: -1
+action_high: 1
+
+num_agents: 1
+random_seed: 2019
+run_name: baseline
+project_name: LunarLanderContinuous
+save_reward_threshold: -100
+use_base_line: 1
+
+policy_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results_LunarLanderContinuous/
+policy_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results_LunarLanderContinuous/policy_best2.pt
+value_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results_LunarLanderContinuous/
+value_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results_LunarLanderContinuous/value_best2.pt
+
+# Training parameters
+
+model: d4pg
+batch_size: 256
+num_steps_train: 300 # number of episodes from all agents
+max_ep_length: 1000 # maximum number of steps per episode
+replay_mem_size: 1000000 # maximum capacity of replay memory
+priority_alpha: 0.6 # controls the randomness vs prioritisation of the prioritised sampling (0.0 = Uniform sampling, 1.0 = Greedy prioritisation)
+priority_beta_start: 0.4 # starting value of beta - controls to what degree IS weights influence the gradient updates to correct for the bias introduces by priority sampling (0 - no correction, 1 - full correction)
+priority_beta_end: 1.0 # beta will be linearly annealed from its start value to this value thoughout training
+discount_rate: 0.99 # Discount rate (gamma) for future rewards
+n_step_returns: 5 # number of future steps to collect experiences for N-step returns
+update_agent_ep: 3 # agent gets latest parameters from learner every update_agent_ep episodes
+replay_queue_size: 64 # queue with replays from all the agents
+batch_queue_size: 64 # queue with batches given to learner
+replay_memory_prioritized: 0 # flag to use prioritized buffer
+device: cuda
+num_episode_save: 500
+
+# Network parameters
+
+critic_learning_rate: 0.0005
+actor_learning_rate: 0.0005
+dense_size: 300 # size of the 2 hidden layers in networks
+final_layer_init: 0.003
+num_atoms: 51 # number of atoms in output layer of distributed critic
+v_min: -100.0 # lower bound of critic value output distribution
+v_max: 100.0 # upper bound of critic value output distribution
+tau: 0.001 # parameter for soft target network updates
+
+# Miscellaneous
+
+results_path: results
