+++ far_ws/src/follow_ahead_rl/scripts/d4pg/configs/follow_rl.yml
@@ -1,19 +1,20 @@
 # Environment parameters
 
 env: gazeboros-v0
-state_dim: 27
+state_dim: 27 # or 47? this should be inferred from a util.py rad_config anyways
 action_dim: 2
 action_low: -1
 action_high: 1
-num_agents: 4
+num_agents: 4 # 4
 random_seed: 2019
 run_name: base_line
 use_base_line: 1
 project_name: follow_ahead_d4pg_v1_auto
-policy_weights: results/
-policy_weights_best: results/policy_best2.pt
-value_weights: results/
-value_weights_best: results/value_best2.pt
+save_reward_threshold : -100
+policy_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/
+policy_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/policy_best2.pt
+value_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/
+value_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/value_best2.pt
 
 # Training parameters
 
@@ -39,7 +40,8 @@ device: cuda
 
 critic_learning_rate: 0.0005
 actor_learning_rate: 0.0005
-dense_size: 400 # size of the 2 hidden layers in networks
+### this is changed by hand in networks.py such that dense_2 is 300
+dense_size: 300 # 400 # size of the 2 hidden layers in networks
 final_layer_init: 0.003
 num_atoms: 51 # number of atoms in output layer of distributed critic
 v_min: -20.0 # lower bound of critic value output distribution
