+++ far_ws/src/follow_ahead_rl/scripts/d4pg/configs/reachability_rl.yml
@@ -6,10 +6,10 @@ action_low: -1
 action_high: 1
 num_agents: 4
 random_seed: 2019
-policy_weights: /home/payam/ros2_ws/src/follow_ahead_rl/scripts/results/
-policy_weights_best: /home/payam/ros2_ws/src/follow_ahead_rl/scripts/results/policy_best2.pt
-value_weights: /home/payam/ros2_ws/src/follow_ahead_rl/scripts/results/
-value_weights_best: /home/payam/ros2_ws/src/follow_ahead_rl/scripts/results/value_best2.pt
+policy_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/
+policy_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/policy_best2.pt
+value_weights: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/
+value_weights_best: /home/alik604/yet-another-friendly-companion/far_ws/src/follow_ahead_rl/scripts/results/value_best2.pt
 
 # Training parameters
