+++ far_ws/src/follow_ahead_rl/scripts/d4pg/models/d4pg/d4pg.py
@@ -12,7 +12,56 @@ from utils.reward_plot import plot_rewards
 from utils.logger import Logger
 
 from .networks import PolicyNetwork, ValueNetwork
-from .l2_projection import _l2_project
+
+def _l2_project(z_p, p, z_q):
+    """Projects distribution (z_p, p) onto support z_q under L2-metric over CDFs.
+    The supports z_p and z_q are specified as tensors of distinct atoms (given
+    in ascending order).
+    Let Kq be len(z_q) and Kp be len(z_p). This projection works for any
+    support z_q, in particular Kq need not be equal to Kp.
+    Args:
+      z_p: Tensor holding support of distribution p, shape `[batch_size, Kp]`.
+      p: Tensor holding probability values p(z_p[i]), shape `[batch_size, Kp]`.
+      z_q: Tensor holding support to project onto, shape `[Kq]`.
+    Returns:
+      Projection of (z_p, p) onto support z_q under Cramer distance.
+    """
+    # Broadcasting of tensors is used extensively in the code below. To avoid
+    # accidental broadcasting along unintended dimensions, tensors are defensively
+    # reshaped to have equal number of dimensions (3) throughout and intended # shapes are indicated alongside tensor definitions. To reduce verbosity,
+    # extra dimensions of size 1 are inserted by indexing with `None` instead of
+    # `tf.expand_dims()` (e.g., `x[:, None, :]` reshapes a tensor of shape
+    # `[k, l]' to one of shape `[k, 1, l]`).
+
+    ## l2_projection ##
+    # Adapted for PyTorch from: https://github.com/deepmind/trfl/blob/master/trfl/dist_value_ops.py
+    # Projects the target distribution onto the support of the original network [Vmin, Vmax]
+
+    z_p = torch.tensor(z_p).float()
+
+    # Extract vmin and vmax and construct helper tensors from z_q
+    vmin, vmax = z_q[0], z_q[-1]
+
+    d_pos = torch.cat([z_q, vmin[None]], 0)[1:]
+    d_neg = torch.cat([vmax[None], z_q], 0)[:-1]
+
+    # Clip z_p to be in new support range (vmin, vmax)
+    z_p = torch.clamp(z_p, vmin, vmax)[:, None, :]
+
+    # Get the distance between atom values in support
+    d_pos = (d_pos - z_q)[None, :, None]
+    d_neg = (z_q - d_neg)[None, :, None]
+    z_q = z_q[None, :, None]
+
+    d_neg = torch.where(d_neg>0, 1./d_neg, torch.zeros(d_neg.shape))
+    d_pos = torch.where(d_pos>0, 1./d_pos, torch.zeros(d_pos.shape))
+
+    delta_qp = z_p - z_q
+    d_sign = (delta_qp >= 0).type(p.dtype)
+
+    delta_hat = (d_sign * delta_qp * d_pos) - ((1. - d_sign) * delta_qp * d_neg)
+    p = p[:, None, :]
+    return torch.sum(torch.clamp(1. - delta_hat, 0., 1.) * p, -1)
 
 
 class LearnerD4PG(object):
