+++ far_ws/src/follow_ahead_rl/scripts/d4pg/test.py
@@ -3,6 +3,7 @@ from utils.utils import read_config
 
 if __name__ == "__main__":
     CONFIG_PATH = "d4pg-pytorch/configs/follow_rl.yml"
+    CONFIG_PATH = "configs/LunarLanderContinuous_d4pg.yml"
     config = read_config(CONFIG_PATH)
 
     engine = load_engine(config)
