+++ far_ws/src/follow_ahead_rl/scripts/d4pg/tests/lunar_lander_continous/test_lunarlander_continous_d4pg.py
@@ -1,5 +1,5 @@
 import unittest
-from models.d3pg.train import train, read_config
+from models.d3pg.train import train, read_config # not a mistake...
 
 
 class TestsLunarLanderContinousD4PG(unittest.TestCase):
