+++ far_ws/src/follow_ahead_rl/scripts/d4pg/train.py
@@ -1,9 +1,12 @@
 from models.engine import load_engine
+from models.d4pg.engine import Engine as D4PGEngine
+
 from utils.utils import read_config
 
 if __name__ == "__main__":
-    CONFIG_PATH = "d4pg-pytorch/configs/follow_rl.yml"
+    #CONFIG_PATH = "configs/LunarLanderContinuous_d4pg.yml"
+    CONFIG_PATH = "configs/follow_rl.yml" # d4pg-pytorch/
     config = read_config(CONFIG_PATH)
 
-    engine = load_engine(config)
+    engine = D4PGEngine(config) #load_engine(config)
     engine.train()
